Task 41 Triangle or not solved

+++ Task041_TriangleOrNot/Program.cs
@@ -1,7 +1,7 @@
 ﻿// 41. Выяснить являются ли три числа сторонами треугольника
 
 // функция для проверки могут ли три отрезка образовать треугольник или нет
-string TriangleOrNot(int a, int b, int c) 
+string TriangleOrNot(double a, double b, double c) 
 {   // Сумма длин двух любых сторон треугольника больше длины оставшейся стороны:
     if ( (a + b > c) && (a + c > b) && (b + c > a) )  
         return $"Введенные три отрезка могут образовать треугольник";
@@ -11,9 +11,9 @@ string TriangleOrNot(int a, int b, int c)
 
 Console.Clear();
 Console.WriteLine("Введите три числа:");
-int a = Convert.ToInt32(Console.ReadLine());
-int b = Convert.ToInt32(Console.ReadLine());
-int c = Convert.ToInt32(Console.ReadLine());
+double a = Convert.ToDouble(Console.ReadLine());
+double b = Convert.ToDouble(Console.ReadLine());
+double c = Convert.ToDouble(Console.ReadLine());
 
 // выводим на экран значение функции при заданных значениях координат
 Console.WriteLine(TriangleOrNot(a, b, c));
